fix debugging stl iterators assert in bibliography toolbar
bibliography: click autofilter toolbar icon to get drop down, select an
entry and assert on use of invalid iterators.

In this specific case we want to reset them all, so add a new simpler
method to just do that and not do the (always supposed to be true) comparison

note: in ORowSet::checkUpdateIterator
at m_aCurrentRow = m_pCache->m_aInsertRow; these iterators are then invalid
the whole iterator caching and fiddling here is concerning.

Change-Id: I74952896f47ac482b6bacb26afe5b1cfcf6bc49c