Resolves: tdf#82781 avoid dereferencing invalid iterators
when using print preview with cursor on 2nd page which consists solely of
hidden text with show hiddle text enabled.

inner mrParentViewShell.Paint can invalidate iterators, as later
detected with the mbNewLayoutDuringPaint flag but we dereference
for the purposes of drawing the shadow before we get there.

easiest thing seems to be to just grab the bits of info we want before the
inner paint occurs.

Change-Id: I09c9b06449db440b7beba7de9e71c239ee64f6ca