Make the new D2D Graphite rendering work also in the OpenGL case
It used to render just black boxes.

The change in DrawTextImpl() semantics from
0c865931734cda957da506379238b36bf9acbf53 was not properly implemented
in the new Graphite code. The return value is not some kind of
"success" indicator, but tells the caller whether to continue the loop
at that level.

We do need to call FillRect() to fill the requested rectangle of the
HDC with white. On the other hand, the call to Clear() is not needed
and in fact makes no text show up.

(I now see that that same code snippet that calls FillRect() is used
in all the DrawTextImpl() implementations, so it should be factored
out to the call site in WinLayout::DrawText().)

Change-Id: If0533ea1edf065b06ae888c6e57c026f447bcf78