sw: restore lost SwMMResultEmailDialog::SendAsHdl_Impl()
Adapt SwMailBodyDialog to not require a wizard at all, and similarly let
SwGreetingsHandler work without a wizard when m_bIsTabPage is false. Do
this by not calling m_pWizard->GetConfigItem() in SwMailBodyDialog and
in the parent class, instead always work with a new m_rConfigItem, and
let the subclasses decide if they obtain that from a wizard (tabpage
case), or from the view (toolbar case).

With this, it's again possible to add an inline message in case the real
mail merge document is an email attachment (odt/pdf/doc case, does not
affect html or plain text).

Change-Id: I4bf68ac4c06772b50d468d6b30959c5970babcb3