Revert "tdf#96393: reuse layout if available"
This reverts commit 93df63bf05588127b0d119324d368a21068e2c15.

Not only introduced it the regression of
https://bugs.documentfoundation.org/show_bug.cgi?id=98437 in that
dialogs with options shown closed in a different size than they would be
opened with options not shown now open in the previous closed size, but
it also introduces regressions for dialogs with nested windows where if
the initial layout is skipped for the remembered outer window size the
inner windows are not initialized anymore.

One example is the Calc Manage Names dialog, where the table is a nested
window that now has an initial width 0 and thus the column headers that
should be equally sized distributed are lumped together at the top left
and only two are displayed instead of three and the table content is
messed up.

Change-Id: I319d931cde2b7b87261e4c04122b91cd0f64b117
Reviewed-on: https://gerrit.libreoffice.org/23402
Reviewed-by: pepp <pelloux@gmail.com>
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>