nss: remove nss-pem.patch
It looks like commit 41f02a72735230928a0704137c800e9b9cd5d3bc made
this obsolete: by default, curl's configure will pick some OpenSSL
certificate database in /etc to use, which is in PEM format.

But since every distribution uses a different path in /etc for that,
it will work only on the same distribution as the build host anyway,
which makes it mostly useless in practice.

So we don't need our NSS to be able to read PEM format now.

Speaking of which, PEM support was only required for Linux builds
anyway, not for WNT/Mac builds.

Change-Id: Ib9bc6ec0b7d71c14e9a53454e6efed038072e913
Reviewed-on: https://gerrit.libreoffice.org/23435
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Giuseppe Castagno <giuseppe.castagno@acca-esse.eu>
Reviewed-by: Michael Stahl <mstahl@redhat.com>