~OfficeIPCThread can only be reached with pGlobalOfficeIPCThread cleared
...as only one instance of OfficeIPCThread is ever created (in
OfficeIPCThread::EnableOfficeIPCThread), which is then either assigned to
pGlobalOfficeIPCThread (if this is the first soffice instance), or
pGlobalOfficeIPCThread is left unassigned and the OfficeIPCThread instance goes
out of scope at the end of EnableOfficeIPCThread (if another soffice instance is
already running).

So mpDispatchWatcher also does not need to be cleared here, as it is only ever
accessed through pGlobalOfficeIPCThread.  And thus, there is no longer a need to
lock GetMutex() here, either.

Change-Id: I56cffc43361d01d747084153f318b8488519cf9b