tdf#88453 sw layout, nested table split: shrink inner tables as well
Commit 3d14f27403ae153469c0bfa9699f2c009d1baa44 (tdf#88453 sw layout:
fix split of nested tables with large amount of rows, 2016-03-23) took
care of two similar problems when splitting tables which has nested
ones (setting cell height to 0 when its content height is 0, and setting
row height to 0 when its cells' height are zero).

Turns out there is a third aspect of the same problem: when row heights
are set to 0 during shrinking for table split purposes, the table height
has to be set to 0 as well, in case the table contains only 0-height
rows. With this, the original bugdoc's layout is the expected one.

Change-Id: I3f5392eb914a60e9caa5182dc6d44bd3a93f41d6
Reviewed-on: https://gerrit.libreoffice.org/23497
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>