tdf#84938 replace #defined constants with enum class
replace if constructs with switch
remove temp variables

tried to squeeze the enum in one of the 3 classes
but it doesn't feel right anywhere.
ScHeaderFooterContentObj doesn't use it, except for initialzing
ScHeaderFooterTextObj.
ScHeaderFooterTextData holds it but nothing more.
ScHeaderFooterTextObj uses it but always has to get it via
ScHeaderFooterTextData.GetPart()

Change-Id: I838ebff2e116e24a1c9359d6652bf5b1d30da48a
Reviewed-on: https://gerrit.libreoffice.org/23352
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>