Avoid running out of memory on Windows when OpenCL is available
Call Clear() on the ScDocShellRef (derived from tools::SvRef) members
of the ScOpenCLTest object in tearDown().

Otherwise the memory used by the ScDocShellRef objects in each
ScOpenCLTest object would be held until when all the ScOpenCLTest
objects are destroyed in the end. (CppUnit creates all the
ScOpenCLTest objects in advance, then calls the test functions on
them, and then destroys them.)

The issue was a side-effect of the (otherwise nice) fix for
tdf#89267.

Change-Id: Ie5c59f2f089386c0abc135136397e1653c975fdc