tdf#99100 DOC import: handle subset of STYLEREF natively
Commit 164dd672119aef63aa88568d1277eb37e83286dc (export 'Chapter' field
type as 'StyleRef' into .doc, 2015-08-21) mapped SwChapterField to
STYLEREF in the DOC export. This field type was handled as a field mark
on import. Instead of always handling it as a field mark, recognize the
case when it's the subset we write and we can handle natively, and in
that case create an SwChapterField again on import.

Leave the complex case unchanged as before and keep using field marks
for that.

Change-Id: Icfa8c4be6538da5e02e2d5071af30a46ccfa712b