Related: tdf#99315 BorderLinePrimitive2D: fix solid line primitive width
Regression from commit 5f158beb27f368564fd02b322edf926847053467
(fdo#75260: Correctly draw double lines for both Writer and Calc.,
2014-03-03), the problem was that when decomposing a double border line
promitive to solid line primitives (which happens for the legacy
"border" type, while both the Calc UI and Excel importer creates
"border-thin" ones), the decomposed inner line's height was larger than
the requested width.

As a result there was no gap between the polygon of the inner and the
outer line of the double border, looking like a non-double border.

The width of the outer border is still incorrect on the screen, though.

Change-Id: Ia9713c315ce8f23e2579b257169798e7c82c0a64
Reviewed-on: https://gerrit.libreoffice.org/24115
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>