gengal is not s script anymore so it is already signed on mac
a09116b1c4628596b9102c224f7b028d2731fa78 removed the script layer
around a gengal.bin
so there is no gengal 'script' anymore and gengal is a binary
that was signed at link already.

remove the exception that force a sign of gengal as a 'script'
which caused an error due to re-signing.

Change-Id: Ie22efceb8e5ab20018be784968487917c6f0a95e
Reviewed-on: https://gerrit.libreoffice.org/24306
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>