Pasting -- from a pdf from a fallback font doesn't give -- as output
Type -- into a writer document, on a machine without Courier installed
set that text to Courier. Export to pdf, select inside e.g. evince and
paste to a terminal. You don't get --

0x2D is mapped from unicode 0x2D and unicode 0xAD in the adobe
encoding. So the latter trumps the earlier one and <ADAD> is pasted
instead of <2D2D>.

Reverse the order that the encoding is mapped from unicode so lower
more "ascii" options are preferred over the higher more "unlikely"
unicode options

Change-Id: I8f251253fca468d269493801e668617a935ee15d