logerrit: disarm the "resubmit" foot-gun
Recently some new developers accidentally used "resubmit", intending to
upload a new revision of their change, but it actually always creates a
new gerrit change.

69b14737e1509612f5adbe4875e5fb9e06ecfd56 claims that "resubmit creates a
new Change-Id for the current change and thus allow to submit changes for
review on release branches that were already reviewed on master", but
that is not actually required with gerrit versions that were in use in
the recent years, gerrit is capable of tracking Change-Id per branch.

Change-Id: I66eda4dbc6f4ff1e235105b3f55041cda88a6b60