tdf#42949: prefer including Reference.hxx over Reference.h...
...and so on for Any.hxx and Sequence.hxx.

This commit follows up on two of my previous commits, which are
91b6a2a7a3361af3c6b666a9b8bd53b0378a7195 and
fbfe54f32ac03c4a424f84df36a23ba4d8d26705.

Many non-static inline functions that are declared in Reference.h and
similar files are not defined in the .h file, but their definitions
are in the corresponding .hxx files instead. Thus, let's prefer
including the .hxx files over the .h one so that we don't include
inline functions without their definitions.

In practice, the corresponding .hxx file is included transitively
from many other places, but it's better not to rely on that.

Change-Id: I3bada39eb879ae736c0360f435b16cd5a61adba8
Reviewed-on: https://gerrit.libreoffice.org/24456
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>