What is bin/distro-install-file-lists supposed to do?
It is called from the recipe for distro-pack-install in Makefile.in, and I'm
(re-/mis-?) using that make target for xdg-app builds (mainly to populate the
/usr/share metadata via bin/distro-install-desktop-integration), which then
fail because there is no $DESTDIR/gid_Module_Root.

Is anybody using that make target still in a way where
bin/distro-install-file-lists is doing something useful (and why is it that
$DESTDIR/gid_Module_Root is existing in that case, but not in mine), or is that
some leftover stuff than should be removed?

Change-Id: I27197c4ae3f57f028f5b181315484e25c1edbaf7